Backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
psycopg[binary]>=3.1.18
pymilvus>=2.3.6
python-dotenv>=1.0.0
pydantic>=2.6.0
ollama>=0.1.6
numpy>=1.26.3